+++ types/era_summary.go
@@ -6,8 +6,8 @@ import "github.com/make-software/casper-go-sdk/types/key"
 type EraSummary struct {
 	// The block hash
 	BlockHash key.Hash `json:"block_hash"`
-	// The Era Id
-	EraID uint32 `json:"era_id"`
+	// The EraID Id
+	EraID uint64 `json:"era_id"`
 	// The StoredValue containing era information.
 	StoredValue StoredValue `json:"stored_value"`
 	// Hex-encoded hash of the state root.
